pyproject.toml
[project]
name = "youtube-video-summarization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.6.0",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
